+++ src/main/scala/is/hail/rvd/OrderedRVDPartitioner.scala
@@ -25,7 +25,7 @@ class OrderedRVDPartitioner(
   require(rangeBounds.isEmpty || (rangeBounds.zip(rangeBounds.tail).forall { case (left: Interval, right: Interval) =>
     !left.mayOverlap(pkType.ordering, right) && pkType.ordering.lteq(left.start, right.start)
   } && rangeBounds.forall { i: Interval =>
-    pkType.ordering.lteq(i.start, i.end) && !i.definitelyEmpty(pkType.ordering)
+    pkType.ordering.lteq(i.start, i.end)
   }))
 
   require(rangeBounds.isEmpty || rangeBounds.zip(rangeBounds.tail).forall { case (left: Interval, right: Interval) =>
